refactor axml, resolve #298
